message→common/internal
