+++ client/meson.build
@@ -53,7 +53,6 @@ webcface_client_lib = static_library('webcface-client',
     filesystem_dep,
     win32folders_dep,
     webcface_common_dep,
-    webcface_message_dep,
     spdlog_dep,
     curl_dep,
   ],
