+++ client/src/canvas2d.cc
@@ -1,7 +1,7 @@
 #include "webcface/canvas2d.h"
 #include "webcface/internal/client_internal.h"
 #include "webcface/member.h"
-#include "webcface/message/message.h"
+#include "webcface/common/internal/message.h"
 #include "webcface/internal/data_buffer.h"
 
 WEBCFACE_NS_BEGIN
