+++ client/src/client_msg_sync.cc
@@ -1,6 +1,6 @@
 #include "webcface/client.h"
 #include "webcface/internal/logger.h"
-#include "webcface/message/message.h"
+#include "webcface/common/internal/message.h"
 #include "webcface/internal/client_internal.h"
 #include "webcface/internal/robot_link_internal.h"
