+++ client/src/func.cc
@@ -1,7 +1,7 @@
 #include "webcface/func.h"
 #include <stdexcept>
 #include "webcface/internal/func_internal.h"
-#include "webcface/message/message.h"
+#include "webcface/common/internal/message.h"
 #include "webcface/internal/client_internal.h"
 #ifdef WEBCFACE_MESON
 #include "webcface-config.h"
