+++ client/src/image.cc
@@ -1,7 +1,7 @@
 #include "webcface/image.h"
 #include "webcface/member.h"
 #include "webcface/internal/client_internal.h"
-#include "webcface/message/message.h"
+#include "webcface/common/internal/message.h"
 #include "webcface/common/encoding.h"
 
 WEBCFACE_NS_BEGIN
