+++ client/src/image_frame.cc
@@ -1,7 +1,7 @@
 #include <cstddef>
 
 #include "webcface/image_frame.h"
-#include "webcface/message/message.h"
+#include "webcface/common/internal/message.h"
 
 WEBCFACE_NS_BEGIN
