+++ client/src/robot_link.cc
@@ -1,5 +1,5 @@
 #include "webcface/robot_link.h"
-#include "webcface/message/message.h"
+#include "webcface/common/internal/message.h"
 #include "webcface/internal/robot_link_internal.h"
 
 WEBCFACE_NS_BEGIN
