+++ client/src/value.cc
@@ -1,7 +1,7 @@
 #include "webcface/internal/client_internal.h"
 #include "webcface/value.h"
 #include "webcface/member.h"
-#include "webcface/message/message.h"
+#include "webcface/common/internal/message.h"
 #include <algorithm>
 #include <cctype>
