+++ client/src/view.cc
@@ -2,7 +2,7 @@
 #include "webcface/internal/client_internal.h"
 #include "webcface/internal/data_store2.h"
 #include "webcface/member.h"
-#include "webcface/message/message.h"
+#include "webcface/common/internal/message.h"
 #include "webcface/internal/data_buffer.h"
 #include "webcface/internal/component_internal.h"
