+++ server-store/meson.build
@@ -27,7 +27,6 @@ webcface_server_store_lib = static_library('webcface-server-store',
     warning_options,
     # webcface_dep,
     webcface_common_dep,
-    webcface_message_dep,
     webcface_client_dep,
     filesystem_dep,
     spdlog_dep,
