+++ tests/dummy_server.cc
@@ -1,4 +1,4 @@
-#include "webcface/message/message.h"
+#include "webcface/common/internal/message.h"
 #include <spdlog/logger.h>
 #include <spdlog/sinks/stdout_color_sinks.h>
 #include <thread>
